move atomic classes from metrics to concurrent classes
