+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/platform/DefaultDoubleGauge.java
@@ -20,8 +20,8 @@ import static com.swirlds.common.metrics.Metric.ValueType.VALUE;
 
 import com.swirlds.base.utility.ToStringBuilder;
 import com.swirlds.common.metrics.DoubleGauge;
-import com.swirlds.common.metrics.atomic.AtomicDouble;
 import com.swirlds.common.metrics.platform.Snapshot.SnapshotEntry;
+import com.swirlds.common.threading.atomic.AtomicDouble;
 import java.util.List;
 
 /**
