+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/metrics/platform/DefaultIntegerPairAccumulator.java
@@ -22,8 +22,8 @@ import com.swirlds.base.utility.ToStringBuilder;
 import com.swirlds.common.metrics.IntegerPairAccumulator;
 import com.swirlds.common.metrics.MetricConfig;
 import com.swirlds.common.metrics.MetricType;
-import com.swirlds.common.metrics.atomic.AtomicIntPair;
 import com.swirlds.common.metrics.platform.Snapshot.SnapshotEntry;
+import com.swirlds.common.threading.atomic.AtomicIntPair;
 import java.util.List;
 import java.util.function.BiFunction;
 import java.util.function.IntSupplier;
