+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/threading/atomic/AtomicIntPair.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.common.metrics.atomic;
+package com.swirlds.common.threading.atomic;
 
 import com.swirlds.common.utility.ByteUtils;
 import java.util.concurrent.atomic.AtomicLong;
