+++ platform-sdk/swirlds-common/src/main/java/module-info.java
@@ -160,6 +160,7 @@ module com.swirlds.common {
             com.swirlds.platform.test,
             com.swirlds.platform.gui;
     exports com.swirlds.common.startup;
+    exports com.swirlds.common.threading.atomic;
 
     requires transitive com.fasterxml.jackson.core;
     requires transitive com.fasterxml.jackson.databind;
