+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/stats/atomic/AtomicSumAndCount.java
@@ -16,7 +16,7 @@
 
 package com.swirlds.platform.stats.atomic;
 
-import com.swirlds.common.metrics.atomic.AtomicIntPair;
+import com.swirlds.common.threading.atomic.AtomicIntPair;
 
 /**
  * Atomically sums the values provided as well as counting the number of additions
